refactor(go/timestamp): move package to timestamp_ext, flatten showcase files, outsource agent executor, and rename fixed time variable
