+++ samples/go/extensions/timestamp/timestamp_ext/client.go
@@ -1,4 +1,4 @@
-package timestamp
+package timestamp_ext
 
 import (
 	"context"
